refactor(monitoring): decompose prediction_log into focused modules
Extract BigQuery operations into _prediction_log_bigquery.py (332 lines)
and shared constants/helpers into _prediction_log_common.py (100 lines).

Reduces prediction_log.py from 826 to 439 lines while maintaining the
same public API. Update test patches to target the correct modules.
